requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.8
django-taggit==6.1.0
dotenv==0.9.9
idna==3.11
pillow==12.0.0
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.3
stripe==14.0.1
typing_extensions==4.15.0
urllib3==2.5.0
